+++ src/interface/ptr.c
@@ -48,7 +48,7 @@ char	*ft_ptrtoa(unsigned long long n, int upper)
 		nb /= 16;
 		len++;
 	}
-	res = (char *)malloc(sizeof(char) * (len + 1));
+	res = (char *)malloc(sizeof(char) * ((unsigned long)len + 1));
 	if (res == NULL)
 		return (NULL);
 	res[len] = '\0';
